cloud-engine/requirement.txt
ArangoPy==0.5.7
boto3==1.7.31
botocore==1.10.31
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-cors-headers==2.3.0
django-filter==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
future==0.16.0
futures==3.2.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
openapi-codec==1.3.2
Pillow==5.1.0
pyArango==1.3.1
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
s3transfer==0.1.13
simplejson==3.15.0
six==1.11.0
slumber==0.7.1
uritemplate==3.0.0
urllib3==1.22
